backend/requirements.txt
requests
pandas
folium
python-dotenv
lxml
plotly
scikit-learn
sqlalchemy
